% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sql_database.R
\name{create_sql_database}
\alias{create_sql_database}
\title{Create genome sql database
This function creates sql databases from provided gff files
(which must have been mapped to the GCA_900303365.2 reference genome).
If a genome is provided (as a DNAStringSet) then this information is added
If a directory is provided, the databases are saved in that location}
\usage{
create_sql_database(gff_files, genome = NULL, save_dir = NULL)
}
\arguments{
\item{gff_files}{a list of gff files from which to create the databases}

\item{genome}{a DNAStringSetcontaining the genome}

\item{save_dir}{An existing directory where the databases will be saved}
}
\value{
a list of sql databases
}
\description{
Create genome sql database
This function creates sql databases from provided gff files
(which must have been mapped to the GCA_900303365.2 reference genome).
If a genome is provided (as a DNAStringSet) then this information is added
If a directory is provided, the databases are saved in that location
}
