% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_GeneRegionTrack.R
\name{gene_GeneRegionTrack}
\alias{gene_GeneRegionTrack}
\title{Creates a gene region track of the selected gene using a model and sql database}
\usage{
gene_GeneRegionTrack(my_gene, sql_database, my_model = NULL)
}
\arguments{
\item{my_gene}{gene name}

\item{sql_database}{A list of sql_databases created by create_sql_database()}

\item{my_model}{the sql database name in which the gene is found}
}
\value{
a gene region track suitable for display with plotTracks
}
\description{
Creates a gene region track of the selected gene using a model and sql database
}
