% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_transcripts.R
\name{get_all_transcripts}
\alias{get_all_transcripts}
\title{Creates a dataframe of all transcript names and corresponding databases}
\usage{
get_all_transcripts(sql_database)
}
\arguments{
\item{sql_database}{A list of sql_databases created by create_sql_database()}
}
\value{
a dataframe of transcript names and databases
}
\description{
Creates a dataframe of all transcript names and corresponding databases
}
