% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_from_gene.R
\name{get_model_from_gene}
\alias{get_model_from_gene}
\title{Gets the gene model name from a gene}
\usage{
get_model_from_gene(my_gene, sql_database)
}
\arguments{
\item{my_gene}{gene name}

\item{sql_database}{A list of sql_databases created by create_sql_database()}
}
\value{
the gene model containing this gene
}
\description{
Gets the gene model name from a gene
}
