% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overlapping_genes.R
\name{get_overlapping_genes}
\alias{get_overlapping_genes}
\title{Creates display tracks for genes that overlap the selected gene}
\usage{
get_overlapping_genes(my_gene, sql_database)
}
\arguments{
\item{my_gene}{gene name}

\item{sql_database}{A list of sql_databases created by create_sql_database()}
}
\value{
display tracks of all genes which overlap
}
\description{
Creates display tracks for genes that overlap the selected gene
}
