% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clubroot_genome.R
\name{load_clubroot_genome}
\alias{load_clubroot_genome}
\title{Load Pb314v2 genome
This function loads the Pb314v2 genome
This is the GCA_900303365.2 reference genome.}
\usage{
load_clubroot_genome(location = NULL)
}
\arguments{
\item{location}{path to directory containing genome fasta file or NULL to use package data}
}
\value{
a DNA string set of the fasta file to which everything has been mapped
}
\description{
Load Pb314v2 genome
This function loads the Pb314v2 genome
This is the GCA_900303365.2 reference genome.
}
