% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gff.R
\name{load_gff}
\alias{load_gff}
\title{Load GFF files
This function loads gff files
(which must have been mapped to the
GCA_900303365.2 reference genome).}
\usage{
load_gff(location = NULL)
}
\arguments{
\item{location}{path to directory containing gff files or NULL to use package data}
}
\value{
a list of gff files
}
\description{
Load GFF files
This function loads gff files
(which must have been mapped to the
GCA_900303365.2 reference genome).
}
