% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_interproscan.R
\name{load_interproscan}
\alias{load_interproscan}
\title{Load Interproscan files files
Interproscan TSV output matched to proteins extracted from gff files}
\usage{
load_interproscan(location = NULL)
}
\arguments{
\item{location}{path to directory containing interproscan files or NULL to use package data (end in .ipro.tsv)}
}
\value{
a list of tsv files
}
\description{
Load Interproscan files files
Interproscan TSV output matched to proteins extracted from gff files
}
