% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sql_database.R
\name{load_sql_database}
\alias{load_sql_database}
\title{Loads sql databases
This function loads sql databases from the provided location}
\usage{
load_sql_database(location)
}
\arguments{
\item{location}{An existing directory where the databases are located}
}
\value{
a list of sql databases
}
\description{
Loads sql databases
This function loads sql databases from the provided location
}
