% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_interproscan.R
\name{map_interproscan}
\alias{map_interproscan}
\title{Map Interproscan tsv to a gene
Interproscan TSV output matched to proteins extracted from gff files}
\usage{
map_interproscan(my_gene, ipro_list, my_genome, sql_database, my_model = NULL)
}
\arguments{
\item{my_gene}{gene to map}

\item{ipro_list}{list of ipro tsv files from load_interproscan()}

\item{my_genome}{Pb314v2 genome from load_genome()}

\item{sql_database}{Sql databases from load_sql_database()}
}
\value{
a dataframe of ipro motifs mapped to the gene
}
\description{
Map Interproscan tsv to a gene
Interproscan TSV output matched to proteins extracted from gff files
}
